+++ source/OffsetGenerator.h
@@ -21,6 +21,8 @@ class OffsetGenerator
 		virtual ~OffsetGenerator() {}
 
 		virtual void reset() = 0; // reset for reuse of object with next file
+		virtual void reset(uint64_t len, uint64_t offset) = 0; /* warning: for random generators,
+													randAmount is always set to range len */
 		virtual uint64_t getNextOffset() = 0;
 		virtual size_t getBlockSize() const = 0; // tyically you want getNextBlockSizeToSubmit()
 		virtual size_t getNextBlockSizeToSubmit() const = 0;
@@ -46,9 +48,9 @@ class OffsetGenSequential : public OffsetGenerator
 		virtual ~OffsetGenSequential() {}
 
 	protected:
-		const uint64_t numBytesTotal;
+		uint64_t numBytesTotal;
 		uint64_t numBytesLeft;
-		const uint64_t startOffset;
+		uint64_t startOffset;
 		uint64_t currentOffset;
 		const size_t blockSize;
 
@@ -58,6 +60,14 @@ class OffsetGenSequential : public OffsetGenerator
 		{
 			numBytesLeft = numBytesTotal;
 			currentOffset = startOffset;
+		}
+
+		virtual void reset(uint64_t len, uint64_t offset) override
+		{
+			numBytesTotal = len;
+			numBytesLeft = len;
+			startOffset = offset;
+			currentOffset = offset;
 		}
 
 		virtual uint64_t getNextOffset() override
@@ -98,9 +108,9 @@ class OffsetGenReverseSeq : public OffsetGenerator
 		virtual ~OffsetGenReverseSeq() {}
 
 	protected:
-		const uint64_t numBytesTotal;
+		uint64_t numBytesTotal;
 		uint64_t numBytesLeft;
-		const uint64_t startOffset;
+		uint64_t startOffset;
 		uint64_t currentOffset;
 		const size_t blockSize;
 
@@ -124,6 +134,16 @@ class OffsetGenReverseSeq : public OffsetGenerator
 				currentOffset = startOffset + numBytesTotal - lastBlockRemainder;
 			else
 				currentOffset = startOffset + numBytesTotal - blockSize;
+		}
+
+		virtual void reset(uint64_t len, uint64_t offset) override
+		{
+			numBytesTotal = len;
+			numBytesLeft = len;
+			startOffset = offset;
+			currentOffset = offset;
+
+			reset();
 		}
 
 		virtual uint64_t getNextOffset() override
@@ -157,19 +177,16 @@ class OffsetGenReverseSeq : public OffsetGenerator
 class OffsetGenRandom : public OffsetGenerator
 {
 	public:
-		OffsetGenRandom(const ProgArgs& progArgs, RandAlgoInterface& randAlgo, uint64_t len,
+		OffsetGenRandom(uint64_t numBytesTotal, RandAlgoInterface& randAlgo, uint64_t len,
 			uint64_t offset, size_t blockSize) :
-			progArgs(progArgs), randRange(randAlgo, offset, offset + len - blockSize),
+			randRange(randAlgo, offset, offset + len - blockSize),
+			numBytesTotal(numBytesTotal), numBytesLeft(numBytesTotal),
 			blockSize(blockSize)
-		{
-			this->numBytesTotal = progArgs.getRandomAmount() / progArgs.getNumDataSetThreads();
-			this->numBytesLeft = this->numBytesTotal;
-		}
+		{ }
 
 		virtual ~OffsetGenRandom() {}
 
 	protected:
-		const ProgArgs& progArgs;
 		RandAlgoRange randRange;
 
 		uint64_t numBytesTotal;
@@ -180,6 +197,14 @@ class OffsetGenRandom : public OffsetGenerator
 	public:
 		virtual void reset() override
 			{ numBytesLeft = numBytesTotal; }
+
+		virtual void reset(uint64_t len, uint64_t offset) override
+		{
+			numBytesTotal = len;
+			numBytesLeft = len;
+
+			randRange.reset(offset, offset + len - blockSize);
+		}
 
 		virtual uint64_t getNextOffset() override
 			{ return randRange.next(); }
@@ -212,14 +237,13 @@ class OffsetGenRandom : public OffsetGenerator
 class OffsetGenRandomAligned : public OffsetGenerator
 {
 	public:
-		OffsetGenRandomAligned(const ProgArgs& progArgs, RandAlgoInterface& randAlgo, uint64_t len,
+		OffsetGenRandomAligned(uint64_t numBytesTotal, RandAlgoInterface& randAlgo, uint64_t len,
 			uint64_t offset, size_t blockSize) :
-			randRange(randAlgo, 0, (offset + len - blockSize) / blockSize),
+			randRange(randAlgo, 0, (len - blockSize) / blockSize),
+			numBytesTotal(numBytesTotal), numBytesLeft(numBytesTotal),
+			offset(offset),
 			blockSize(blockSize)
-		{
-			this->numBytesTotal = progArgs.getRandomAmount() / progArgs.getNumDataSetThreads();
-			this->numBytesLeft = this->numBytesTotal;
-		}
+		{ }
 
 		virtual ~OffsetGenRandomAligned() {}
 
@@ -228,6 +252,7 @@ class OffsetGenRandomAligned : public OffsetGenerator
 
 		uint64_t numBytesTotal;
 		uint64_t numBytesLeft;
+		uint64_t offset;
 		const size_t blockSize;
 
 		// inliners
@@ -235,8 +260,17 @@ class OffsetGenRandomAligned : public OffsetGenerator
 		virtual void reset() override
 			{ numBytesLeft = numBytesTotal; }
 
+		virtual void reset(uint64_t len, uint64_t offset) override
+		{
+			this->numBytesTotal = len;
+			this->numBytesLeft = len;
+			this->offset = offset;
+
+			randRange.reset(0, (len - blockSize) / blockSize);
+		}
+
 		virtual uint64_t getNextOffset() override
-			{ return randRange.next() * blockSize; }
+			{ return offset + (randRange.next() * blockSize); }
 
 		virtual size_t getBlockSize() const override
 			{ return blockSize; }