+++ source/toolkits/random/RandAlgoXoshiro256ss.h
@@ -18,8 +18,10 @@ class RandAlgoXoshiro256ss : public RandAlgoInterface
 		RandAlgoXoshiro256ss()
 		{
 			// init the 4 state values with random numbers
+			// (32-shift in case random_device returns only 32bit values)
 			for(int i=0; i < 4; i++)
-				state.s[i] = std::random_device()();
+				state.s[i] =
+					( (uint64_t)std::random_device()() << 32) | (uint32_t)std::random_device()();
 		}
 
 		virtual ~RandAlgoXoshiro256ss() {}